@@ -58,40 +58,42 @@
     "redux-react-firebase"
   ],
   "dependencies": {
-    "hoist-non-react-statics": "^2.3.1",
+    "hoist-non-react-statics": "^3.2.1",
     "lodash": "^4.17.11",
     "prop-types": "^15.6.2"
   },
   "peerDependencies": {
-    "react": "^0.14.6 || ^15.0.0-0 || ^16.0.0-0"
+    "react": "^16.4.0"
   },
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.0",
-    "babel-eslint": "^8.0.1",
-    "babel-loader": "^7.1.2",
-    "babel-plugin-add-module-exports": "^0.2.1",
-    "babel-plugin-lodash": "^3.2.11",
+    "@babel/cli": "^7.2.0",
+    "@babel/core": "^7.2.0",
+    "@babel/plugin-proposal-class-properties": "^7.2.1",
+    "@babel/plugin-transform-async-to-generator": "^7.1.0",
+    "@babel/plugin-transform-modules-commonjs": "^7.1.0",
+    "@babel/plugin-transform-regenerator": "^7.0.0",
+    "@babel/plugin-transform-runtime": "^7.2.0",
+    "@babel/preset-env": "^7.2.0",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/register": "^7.0.0",
+    "@babel/runtime": "^7.2.0",
+    "babel-eslint": "^10.0.1",
+    "babel-loader": "^8.0.4",
+    "babel-plugin-add-module-exports": "^1.0.0",
+    "babel-plugin-lodash": "^3.3.4",
     "babel-plugin-module-resolver": "^3.1.1",
-    "babel-plugin-transform-async-to-generator": "^6.24.1",
-    "babel-plugin-transform-class-properties": "^6.24.0",
-    "babel-plugin-transform-export-extensions": "^6.22.0",
-    "babel-plugin-transform-object-assign": "^6.22.0",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-env": "^1.6.0",
-    "babel-preset-react": "^6.23.0",
-    "babel-register": "^6.24.0",
+    "babel-preset-minify": "^0.5.0",
     "chai": "^3.5.0",
-    "chai-as-promised": "^6.0.0",
+    "chai-as-promised": "^7.1.1",
     "chai-enzyme": "^0.8.0",
     "child-process-promise": "^2.2.1",
     "codecov": "^3.1.0",
-    "cross-env": "^5.1.0",
+    "cross-env": "^5.2.0",
     "docdown": "^0.7.2",
     "documentation": "^5.3.2",
     "documentation-markdown-api-theme": "^1.0.2",
-    "enzyme": "^2.3.0",
+    "enzyme": "^2.9.1",
+    "enzyme-adapter-react-16": "^1.7.0",
     "eslint": "^4.8.0",
     "eslint-config-prettier": "^2.3.0",
     "eslint-config-standard": "^10.2.1",
@@ -106,21 +108,22 @@
     "firebase": "4.12.0",
     "firebase-server": "^0.12.0",
     "gitbook-cli": "^2.3.2",
-    "husky": "^0.14.3",
+    "husky": "^1.2.0",
     "istanbul": "1.1.0-alpha.1",
     "jsdom": "^11.10.0",
     "mocha": "^3.5.3",
     "prettier": "1.10.2",
-    "react": "^15.4.2",
-    "react-addons-test-utils": "^15.4.2",
-    "react-dom": "^15.4.2",
-    "recompose": "^0.26.0",
+    "react": "^16.6.3",
+    "react-dom": "^16.6.3",
+    "react-redux": "^6.0.0",
+    "recompose": "^0.30.0",
     "redux": "4.0.0",
     "redux-firestore": "0.5.7",
     "rimraf": "^2.6.2",
     "sinon": "^3.3.0",
-    "sinon-chai": "^2.14.0",
-    "webpack": "^3.6.0",
+    "sinon-chai": "^3.3.0",
+    "webpack": "^4.27.0",
+    "webpack-cli": "^3.1.2",
     "webpack-bundle-analyzer": "^2.9.0",
     "ws": "^3.2.0",
     "xmlhttprequest": "^1.8.0"
