@@ -1,27 +1,9 @@
 import { createStore, compose } from 'redux'
 import rootReducer from './reducer'
-import { firebase as fbConfig } from './config'
-import firebase from 'firebase/app'
-import 'firebase/auth'
-import 'firebase/database'
-import 'firebase/firestore' // make sure you add this for firestore
-import { reactReduxFirebase } from 'react-redux-firebase'
-import { reduxFirestore } from 'redux-firestore'
 
 export default function configureStore (initialState, history) {
-  // Initialize Firebase instance
-  firebase.initializeApp(fbConfig)
-
-  const reduxFirebaseConfig = {
-    userProfile: 'users',
-    useFirestoreForProfile: true,
-    enableLogging: false
-  }
-
   const createStoreWithMiddleware = compose(
-    reactReduxFirebase(firebase, reduxFirebaseConfig),
-    reduxFirestore(firebase),
-    typeof window === 'object' && typeof window.devToolsExtension !== 'undefined' ? window.devToolsExtension() : f => f
+    typeof window === 'object' && typeof window.devToolsExtension !== 'undefined' ? () => window.__REDUX_DEVTOOLS_EXTENSION__ : f => f
   )(createStore)
   const store = createStoreWithMiddleware(rootReducer)
 
