@@ -1,3 +1,4 @@
+import React from 'react'
 // import * as firebase from 'firebase/app.d'
 //* * Declaration file generated by dts-gen */
 
@@ -181,9 +182,29 @@ export function populate(
 
 export function reactReduxFirebase(instance: object, otherConfig: any): any
 
-export function reduxFirebase(instance: any, otherConfig: any): any
+/**
+ * Props passed to ReactReduFirebaseContext component
+ */
+export interface ReactReduxFirebaseContextProps {
+  firebase: object,
+  config: object,
+  dispatch: (action) => void
+  createFirestoreInstance?: (firebase: object, config: object, dispatch: (action) => void) => object
+}
 
-export function reduxReactFirebase(instance: object, otherConfig: any): any
+/**
+ * React Context provider for Firebase instance. Needed to use HOCs like firebaseConnect and withFirebase
+ */
+export namespace ReactReduxFirebaseContext {
+  const prototype: {}
+}
+
+/**
+ * React Context provider for Firebase instance. Needed to use HOCs like firebaseConnect and withFirebase
+ */
+export namespace ReduxFirestoreContext {
+  const prototype: {}
+}
 
 /**
  * React Higher Order Component that passes firebase as a prop (comes from context.store.firebase)
