@@ -2,25 +2,15 @@ import React, { Children, Component, cloneElement } from 'react'
 import PropTypes from 'prop-types'
 import { createSink } from 'recompose'
 import { isObject } from 'lodash'
-import { createStore, compose, combineReducers } from 'redux'
-import { reduxFirestore, reducer as firestoreReducer } from 'redux-firestore'
-import reactReduxFirebase from '../src/enhancer'
+import { createStore, combineReducers } from 'redux'
+import { reducer as firestoreReducer } from 'redux-firestore'
 
 export const storeWithFirebase = () => {
-  const createStoreWithMiddleware = compose(
-    reactReduxFirebase(Firebase, { userProfile: 'users' })
-  )(createStore)
-  return createStoreWithMiddleware(
-    combineReducers({ test: (state = {}) => state })
-  )
+  return createStore(combineReducers({ test: (state = {}) => state }))
 }
 
 export const storeWithFirestore = () => {
-  const createStoreWithMiddleware = compose(
-    reactReduxFirebase(Firebase, { userProfile: 'users' }),
-    reduxFirestore(Firebase) // mock for reduxFirestore from redux-firestore
-  )(createStore)
-  return createStoreWithMiddleware(
+  return createStore(
     combineReducers({
       test: (state = {}) => state,
       firestore: firestoreReducer
